update nightly ut and update neuralchat int8 model loading
