+++ examples/huggingface/pytorch/text-generation/quantization/run_benchmark.sh
@@ -78,7 +78,6 @@ function run_benchmark {
 
 
     if [ "${topology}" = "gpt_j" ]; then
-	pip install transformers==4.31.0
         model_name_or_path="/tf_dataset2/models/pytorch/gpt-j-6B"
     elif [ "${topology}" = "gpt_j_woq_rtn" ]; then
         model_name_or_path="/tf_dataset2/models/pytorch/gpt-j-6B"
@@ -91,13 +90,10 @@ function run_benchmark {
     elif [ "${topology}" = "gpt_j_woq_load8bit" ]; then
         model_name_or_path="/tf_dataset2/models/pytorch/gpt-j-6B"
     elif [ "${topology}" = "opt_1.3b" ]; then
-        pip install transformers==4.31.0
         model_name_or_path="facebook/opt-1.3b"
     elif [ "${topology}" = "opt_2.7b" ]; then
-	pip install transformers==4.31.0
         model_name_or_path="facebook/opt-2.7b"
     elif [ "${topology}" = "opt_6.7b" ]; then
-	pip install transformers==4.31.0
         model_name_or_path="facebook/opt-6.7b"
     elif [ "${topology}" = "bloom_7b1" ]; then
         model_name_or_path="/tf_dataset2/models/pytorch/bloom-7b1"
@@ -106,10 +102,8 @@ function run_benchmark {
     elif [ "${topology}" = "bloomz-3b" ]; then
         model_name_or_path="bigscience/bloomz-3b"
     elif [ "${topology}" = "llama_7b" ]; then
-	pip install transformers==4.31.0
         model_name_or_path="meta-llama/Llama-2-7b-chat-hf"
     elif [ "${topology}" = "llama_13b" ]; then
-	pip install transformers==4.31.0
         model_name_or_path="meta-llama/Llama-2-13b-chat-hf"
     elif [ "${topology}" = "dolly_v2_3b" ]; then
         model_name_or_path="/tf_dataset2/models/pytorch/dolly_v2_3b"
